# docs/v2x-app-api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "V2X MEC Partner API",
    "description": "API for Vehicle-to-Everything (V2X) Multi-Access Edge Computing (MEC) services. This Partner API provides endpoints for ETX registration, V2X Message Deposit, and authentication services.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/prd/v2/configurations/geofence": {
      "get": {
        "tags": [
          "configuration-rest-controller"
        ],
        "summary": "Get geofence by ID",
        "description": "Retrieves detailed information for a specific geofence by its ID. This endpoint returns the complete geofence configuration including geometry, properties, and metadata.",
        "operationId": "getGeofence",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Unique identifier of the geofence",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "geofence-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Geofence details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationGeofenceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid geofence ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication",
            "content": {
              "application/json": {}
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not found - geofence not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "configuration-rest-controller"
        ],
        "summary": "Update existing geofence",
        "description": "Updates an existing geofence with new configuration. All fields in the request will replace the existing geofence configuration.",
        "operationId": "updateGeofence",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Unique identifier of the geofence to update",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "geofence-001"
          }
        ],
        "requestBody": {
          "description": "Updated geofence configuration",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationGeofence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Geofence updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationGeofenceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid geofence configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication",
            "content": {
              "application/json": {}
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not found - geofence not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "configuration-rest-controller"
        ],
        "summary": "Create new geofence",
        "description": "Creates a new geofence with the specified configuration. The geofence can be used for V2X traffic management and safety applications.",
        "operationId": "createGeofence",
        "requestBody": {
          "description": "Geofence configuration to create",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationGeofence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Geofence created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationGeofenceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid geofence configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication",
            "content": {
              "application/json": {}
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "configuration-rest-controller"
        ],
        "summary": "Delete geofence",
        "description": "Deletes a geofence from the system. This operation is irreversible and will remove all associated configuration and data.",
        "operationId": "deleteGeofence",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Unique identifier of the geofence to delete",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "geofence-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Geofence deleted successfully"
          },
          "400": {
            "description": "Bad request - invalid geofence ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication",
            "content": {
              "application/json": {}
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not found - geofence not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/prd/v2/deposit/geofence": {
      "post": {
        "tags": [
          "Deposit"
        ],
        "summary": "Deposit V2X Message",
        "description": "Deposits V2X message data to the database. This endpoint is used to deploy V2X messages (TIM, MAP, etc.) with geospatial information. The configuration must include geofence overrides to define the deployment area. Vendor ID is automatically determined from the user\u0027s JWT token based on their role.",
        "operationId": "deposit",
        "requestBody": {
          "description": "Configuration data to deposit",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "V2X message deployed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeofenceDeploymentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid configuration data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication",
            "content": {
              "application/json": {}
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions",
            "content": {
              "application/json": {}
            }
          },
          "409": {
            "description": "Conflict - no available geohash or conflicting state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large - exceeds maximum allowed geohashes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity - invalid ASN.1 data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Deposit"
        ],
        "summary": "Delete geofence by identifier",
        "description": "Deletes geofence deployments associated with a specific identifier. This endpoint allows for manual removal of geofence deployments by their identifier. The system will deactivate the geofence deployment by setting is_active to false. This follows REST best practices by using query parameters for resource identification.",
        "operationId": "deleteGeofence_1",
        "parameters": [
          {
            "name": "identifier",
            "in": "query",
            "description": "Geofence identifier to delete",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "GEOFENCE_12345"
          }
        ],
        "responses": {
          "204": {
            "description": "Geofence deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication",
            "content": {
              "application/json": {}
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not found - no geofence found with the specified identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/prd/v2/configurations/clear": {
      "post": {
        "tags": [
          "configuration-rest-controller"
        ],
        "summary": "Clear multiple geofences",
        "description": "Clears deployed geofences from the ETX. If set to only TIM messages it will clear all inactive TIMs.This operation can be used to remove multiple geofences at once.",
        "operationId": "clearGeofences",
        "requestBody": {
          "description": "Criteria for clearing geofences",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationClearGeofence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Geofences cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid clear criteria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication",
            "content": {
              "application/json": {}
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/prd/v2/configurations/geofence/ids": {
      "get": {
        "tags": [
          "configuration-rest-controller"
        ],
        "summary": "Get all geofence IDs",
        "description": "Retrieves a list of all available geofence IDs in the ETX. This endpoint returns summary information for all configured geofences in the ETX without the full geofence details.",
        "operationId": "getGeofences",
        "responses": {
          "200": {
            "description": "List of geofence summaries retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationGeofenceSummary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication",
            "content": {
              "application/json": {}
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ConfigurationGeofence": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "geoFence": {
            "$ref": "#/components/schemas/GeofenceFeatureCollection"
          },
          "messageStandard": {
            "type": "string",
            "enum": [
              "etsi",
              "sae"
            ]
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "ConfigurationGeofenceResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "geoFence": {
            "$ref": "#/components/schemas/GeofenceFeatureCollection"
          },
          "messageStandard": {
            "type": "string",
            "enum": [
              "etsi",
              "sae"
            ]
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "vendorId": {
            "type": "string"
          }
        }
      },
...
      "DepositRequest": {
        "type": "object",
        "description": "Request model for V2X message deposit",
        "properties": {
          "asn1_hex": {
            "type": "string",
            "description": "UPER-encoded MessageFrame containing a MAP or TIM V2X Message in hexadecimal format"
          },
          "override_geofence": {
            "$ref": "#/components/schemas/GeofenceFeatureCollection",
            "description": "Override Geofence for ETX deployment. If not set, the geofence will be extracted from the provided message"
          }
        },
        "required": [
          "asn1_hex"
        ]
      },
...
      "ConfigurationClearGeofence": {
        "type": "object",
        "properties": {
          "clear_tim_only": {
            "type": "boolean"
          }
        }
      },
...
      "ConfigurationGeofenceSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "Keycloak JWT Bearer Token. Required for all endpoints except decode endpoints. Obtain a token from the /auth/token endpoint using your credentials.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}